added test for azure oai and updated feature matrix
